requirements.txt
fastapi
uvicorn
python-dotenv
pdfplumber
groq
python-multipart
